requirements.txt
asgiref==3.7.2
dj-database-url==0.5.0
Django==3.2.23
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pytz==2023.3.post1
sqlparse==0.4.4
whitenoise==6.6.0